% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfExport.R
\name{.pdfExport}
\alias{.pdfExport}
\title{.pdfExport}
\usage{
.pdfExport(FileName, date = TRUE)
}
\arguments{
\item{FileName}{Name of the .pdf file you export}

\item{date}{If date = TRUE, it will automatically add date to your file's name}
}
\description{
Open a new .pdf and right in it. IMPORTANT : Set your working directory and use the function dev.off after your code.
}
\examples{
\dontrun{
.pdfExport("the Devil on the night", date = TRUE)
}

}
