% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxExport.R
\name{.xlsxExport}
\alias{.xlsxExport}
\title{.xlsxExport}
\usage{
.xlsxExport(TableExport, FileName, SheetName, date = TRUE)
}
\arguments{
\item{TableExport}{Name of the data frame you wan't to export as .xlsx file}

\item{FileName}{Name of the file you export.}

\item{SheetName}{Name of the sheet in the .xlsx file}

\item{date}{If date = TRUE, date will automatically write in the FileName}
}
\description{
Export a data.frame as an .xlsx file. Don't forget to set wour working directory before using it.
}
\examples{
\dontrun{
.xlsxExport(data,"data from devil","Devils", date = TRUE)}

}
